Refactor the add album method and musicalbum class
