+++ lib/records/items.json
@@ -1,1 +1,1 @@
-[]
+[{"type":"MusicAlbum","id":9433,"genre_name":"name","publish_date":0,"on_spotify":false},{"type":"MusicAlbum","id":58,"genre_name":"asd","publish_date":2023,"on_spotify":true}]
